refactor(admin): AdminConnections.tsx 797줄을 6 파일로 분해
`pages/admin/AdminConnections.tsx` 한 컴포넌트에 IAM source 목록 +
5종 provider form + SSO 상태 + AI provider 4종 등록 + Webhook 안내가
다 박혀 있었음. 책임 단위로 쪼개 main을 orchestrator로 줄였음.

pages/admin/connections/
  constants.ts            KIND_LABEL / STATUS_* / AI_PROVIDERS 메타
  IAMSourceCard.tsx       Source table + sync 액션
  IAMSourceModal.tsx      신규 source 추가 modal +
                          AWS/K8s/LDAP/GCP/Azure/Custom 6 form fields
  SSOCard.tsx             SSO 모드 / 활성 IdP / .env 힌트
  AIProvidersCard.tsx     AI provider 4종 등록 · 활성화 · 테스트 · 삭제
  WebhookCard.tsx         GitHub webhook 안내

AdminConnections.tsx (main): 797줄 → 44줄. import + state +
5개 카드 배치만.

호환성: default export 그대로. 라우터 측 import 변경 0건.

검증:
- tsc --noEmit 통과
- /admin/connections 페이지 200 OK
- 5개 IAM provider form 모두 노출
- AI provider 4종 카드 + 등록 modal 정상 (test connection 포함)
